Добавил компонент Хранилища сертификатов и перенес в него инициализацию
