+++ src/eos-parameters/_styles.scss
@@ -1,6 +1,7 @@
 @import "./parametersSystem/parametersSystem.component";
 @import "./parametersSystem/shared/param-header/header.component";
 @import "./parametersSystem/param-web/param-web.component";
+@import "./parametersSystem/param-web/cert-stores/cert-stores.component";
 @import "./parametersSystem/param-other/param-other.component";
 @import "./parametersSystem/param-search/param-search.component";
 @import "./parametersSystem/param-context-rc/param-context-rc.component";
