+++ src/eos-parameters/eos-parameters.module.ts
@@ -20,6 +20,7 @@ import { AuthenticationCollectionComponent } from './parametersSystem/param-auth
 import { ParamFielsComponent } from './parametersSystem/param-files/param-files.component';
 import { ParamPrjRcComponent } from './parametersSystem/param-prj-rc/param-prj-rc.component';
 import { ParamRcComponent } from './parametersSystem/param-rc/param-rc.component';
+import { CertStoresComponent } from './parametersSystem/param-web/cert-stores/cert-stores.component';
 
 
 
@@ -41,7 +42,8 @@ import { ParamDescriptorSrv } from './parametersSystem/shared/service/param-desc
         AuthenticationCollectionComponent,
         ParamFielsComponent,
         ParamPrjRcComponent,
-        ParamRcComponent
+        ParamRcComponent,
+        CertStoresComponent,
     ],
     imports: [
         BrowserModule,
