+++ src/eos-parameters/parametersSystem/param-web/cert-stores/cert-stores.component.html
@@ -1,0 +1,4 @@
+<span class="l-font">Хранилища сертификатов</span>
+
+<pre>{{carmaStores | json}}</pre>
+
